# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: AI Career Guidance API for pre-university students in Zimbabwe
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /recommend:
    post:
      operationId: getRecommendations
      tags: [careers]
      summary: Get personalized career recommendations based on student profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentProfile"
      responses:
        "200":
          description: Career recommendations with matched programs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecommendationsResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /careers/{careerId}:
    get:
      operationId: getCareerById
      tags: [careers]
      summary: Get detailed information about a specific career
      parameters:
        - name: careerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Career details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Career"
        "404":
          description: Career not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /feedback:
    post:
      operationId: submitFeedback
      tags: [feedback]
      summary: Submit feedback about career recommendations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedbackRequest"
      responses:
        "201":
          description: Feedback submitted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Career:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        category:
          type: string
        requiredSkills:
          type: array
          items:
            type: string
        aLevelSubjects:
          type: array
          items:
            type: string
        universityPrograms:
          type: array
          items:
            type: string
        averageSalary:
          type: string
        jobOutlook:
          type: string
      required: [id, name, description, category, requiredSkills, aLevelSubjects, universityPrograms, averageSalary, jobOutlook]
    StudentProfile:
      type: object
      properties:
        interests:
          type: array
          items:
            type: string
        strengths:
          type: array
          items:
            type: string
        subjects:
          type: array
          items:
            type: string
          description: A-Level subjects the student has passed
        oLevelSubjects:
          type: array
          items:
            type: string
          description: O-Level subjects the student has passed
        personalityType:
          type: ["string", "null"]
        hobbies:
          type: array
          items:
            type: string
        cutOffPoints:
          type: ["integer", "null"]
          description: Student's ZIMSEC cut-off points (optional)
        oLevelPasses:
          type: ["integer", "null"]
          description: Number of O-Level passes (minimum 5 required for most programs)
        aLevelPasses:
          type: ["integer", "null"]
          description: Number of A-Level passes (minimum 2 required for most programs)
      required: [interests, strengths, subjects]
    AiRecommendedProgram:
      type: object
      properties:
        programName:
          type: string
        schoolName:
          type: string
      required: [programName, schoolName]
      description: Program AI recommended, verified against DB
    RecommendationsResponse:
      type: object
      properties:
        recommendations:
          type: array
          items:
            $ref: "#/components/schemas/CareerRecommendation"
        aiAdvice:
          type: string
          description: AI-generated personalized advice based on profile and DB programs
        aiRecommendedPrograms:
          type: array
          items:
            $ref: "#/components/schemas/AiRecommendedProgram"
          description: Programs AI recommended that exist in our database (cross-referenced)
      required: [recommendations]
    CareerRecommendation:
      type: object
      properties:
        career:
          $ref: "#/components/schemas/Career"
        matchPercentage:
          type: integer
        matchReasons:
          type: array
          items:
            type: string
        demandLevel:
          type: string
          enum: [High, Medium, Low]
        matchedPrograms:
          type: array
          items:
            $ref: "#/components/schemas/ProgramMatch"
          description: University programs the student qualifies for based on their subjects
      required: [career, matchPercentage, matchReasons, demandLevel, matchedPrograms]
    UniversityProgram:
      type: object
      properties:
        id:
          type: integer
        schoolName:
          type: string
        programName:
          type: string
        faculty:
          type: ["string", "null"]
        requiredSubjects:
          type: array
          items:
            type: string
          description: Required A-Level subjects
        minRequiredSubjects:
          type: ["integer", "null"]
          description: Minimum number of required subjects needed (e.g. 2 = at least 2 of the listed). Null = all required.
        minimumPoints:
          type: ["integer", "null"]
          description: Minimum ZIMSEC cut-off points required
        minOLevelPasses:
          type: ["integer", "null"]
          description: Minimum O-Level passes required (default 5)
        minALevelPasses:
          type: ["integer", "null"]
          description: Minimum A-Level passes required (default 2)
        duration:
          type: ["string", "null"]
          description: Duration of program e.g. 4 years
        description:
          type: ["string", "null"]
        careerCategory:
          type: ["string", "null"]
          description: Related career category
      required: [id, schoolName, programName, requiredSubjects]
...
    ProgramMatch:
      type: object
      properties:
        program:
          $ref: "#/components/schemas/UniversityProgram"
        qualifies:
          type: boolean
          description: Whether the student fully qualifies
        missingSubjects:
          type: array
          items:
            type: string
          description: Subjects the student is missing
        meetsPointsRequirement:
          type: ["boolean", "null"]
          description: Whether student meets the cut-off points (null if no points provided or required)
        meetsOLevelRequirement:
          type: ["boolean", "null"]
          description: Whether student meets O-Level pass requirement
        meetsALevelRequirement:
          type: ["boolean", "null"]
          description: Whether student meets A-Level pass requirement
        pointsChance:
          type: ["string", "null"]
          enum: [high, equal, low]
          description: With user's points - high/equal/low chance to enroll (informational only, does not affect qualification)
      required: [program, qualifies, missingSubjects]
...
    FeedbackRequest:
      type: object
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: ["string", "null"]
        careerName:
          type: ["string", "null"]
        helpful:
          type: boolean
      required: [rating, helpful]
    FeedbackResponse:
      type: object
      properties:
        id:
          type: integer
        message:
          type: string
      required: [id, message]