# lib/api-zod/src/generated/types/recommendationsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Career Guidance API for pre-university students in Zimbabwe
 * OpenAPI spec version: 0.2.0
 */
import type { AiRecommendedProgram } from "./aiRecommendedProgram";
import type { CareerRecommendation } from "./careerRecommendation";

export interface RecommendationsResponse {
  recommendations: CareerRecommendation[];
  /** AI-generated personalized advice based on profile and DB programs */
  aiAdvice?: string;
  /** Programs AI recommended that exist in our database (cross-referenced) */
  aiRecommendedPrograms?: AiRecommendedProgram[];
}