# lib/api-zod/src/generated/types/universityProgram.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Career Guidance API for pre-university students in Zimbabwe
 * OpenAPI spec version: 0.2.0
 */

export interface UniversityProgram {
  id: number;
  schoolName: string;
  programName: string;
  /** @nullable */
  faculty?: string | null;
  /** Required A-Level subjects */
  requiredSubjects: string[];
  /**
   * Minimum number of required subjects needed (e.g. 2 = at least 2 of the listed). Null = all required.
   * @nullable
   */
  minRequiredSubjects?: number | null;
  /**
   * Minimum ZIMSEC cut-off points required
   * @nullable
   */
  minimumPoints?: number | null;
  /**
   * Minimum O-Level passes required (default 5)
   * @nullable
   */
  minOLevelPasses?: number | null;
  /**
   * Minimum A-Level passes required (default 2)
   * @nullable
   */
  minALevelPasses?: number | null;
  /**
   * Duration of program e.g. 4 years
   * @nullable
   */
  duration?: string | null;
  /** @nullable */
  description?: string | null;
  /**
   * Related career category
   * @nullable
   */
  careerCategory?: string | null;
}